update file names and type to user (was teacher)
