+++ api-types/index.ts
@@ -1,1 +1,1 @@
-export * from "./teacher";
+export * from "./user";
