+++ api-types/user.ts
@@ -1,4 +1,4 @@
-export type Teacher = {
+export type User = {
   _id: string;
   firstName: string;
   lastName: string;
