+++ backend/index.ts
@@ -1,11 +1,11 @@
-import { bodyParser } from "@koa/bodyparser"; 
+import { bodyParser } from "@koa/bodyparser";
 import cors from "@koa/cors";
 import Router from "@koa/router";
 import { createServer } from "http";
 import Koa from "koa";
 import serveStatic from "koa-static";
 import { initSocketIo } from "./io";
-import teacherRoutes from "./routes/teachers";
+import teacherRoutes from "./routes/users";
 
 const app = new Koa();
 const router = new Router();
